Redundant Casts

+++ src/main/java/net/minecraft/block/BlockCactus.java
@@ -38,7 +38,7 @@ public class BlockCactus extends Block {
             }
 
             if (i < 3) {
-                int j = ((Integer) state.getValue(AGE)).intValue();
+                int j = state.getValue(AGE).intValue();
 
                 if (j == 15) {
                     worldIn.setBlockState(blockpos, this.getDefaultState());
@@ -104,7 +104,7 @@ public class BlockCactus extends Block {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return ((Integer) state.getValue(AGE)).intValue();
+        return state.getValue(AGE).intValue();
     }
 
     protected BlockState createBlockState() {
