+++ src/main/java/net/minecraft/block/BlockFire.java
@@ -139,7 +139,7 @@ public class BlockFire extends Block {
             if (!flag && worldIn.isRaining() && this.canDie(worldIn, pos)) {
                 worldIn.setBlockToAir(pos);
             } else {
-                int i = ((Integer) state.getValue(AGE)).intValue();
+                int i = state.getValue(AGE).intValue();
 
                 if (i < 15) {
                     state = state.withProperty(AGE, Integer.valueOf(i + rand.nextInt(3) / 2));
@@ -224,12 +224,12 @@ public class BlockFire extends Block {
     }
 
     private int getFlammability(Block blockIn) {
-        Integer integer = (Integer) this.flammabilities.get(blockIn);
+        Integer integer = this.flammabilities.get(blockIn);
         return integer == null ? 0 : integer.intValue();
     }
 
     private int getEncouragement(Block blockIn) {
-        Integer integer = (Integer) this.encouragements.get(blockIn);
+        Integer integer = this.encouragements.get(blockIn);
         return integer == null ? 0 : integer.intValue();
     }
 
@@ -382,7 +382,7 @@ public class BlockFire extends Block {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return ((Integer) state.getValue(AGE)).intValue();
+        return state.getValue(AGE).intValue();
     }
 
     protected BlockState createBlockState() {
