+++ src/main/java/net/minecraft/item/ItemSnow.java
@@ -33,7 +33,7 @@ public class ItemSnow extends ItemBlock {
             }
 
             if (block == this.block) {
-                int i = ((Integer) iblockstate.getValue(BlockSnow.LAYERS)).intValue();
+                int i = iblockstate.getValue(BlockSnow.LAYERS).intValue();
 
                 if (i <= 7) {
                     IBlockState iblockstate1 = iblockstate.withProperty(BlockSnow.LAYERS, Integer.valueOf(i + 1));
